Fix format warnings on x86

+++ hw/xwin/glx/indirect.c
@@ -1146,13 +1146,13 @@ glxWinDeferredCreateDrawable(__GLXWinDrawable *draw, __GLXconfig *config)
 
 #define RASTERWIDTHBYTES(bmi) (((((bmi)->biWidth*(bmi)->biBitCount)+31)&~31)>>3)
             size = bmpHeader.biHeight * RASTERWIDTHBYTES(&bmpHeader);
-            ErrorF("shared memory region size %zu + %u\n", sizeof(BITMAPINFOHEADER), size);
+            ErrorF("shared memory region size %zu + %u\n", sizeof(BITMAPINFOHEADER), (unsigned int)size);
 
             // Create unique name for mapping based on XID
             //
             // XXX: not quite unique as potentially this name could be used in
             // another server instance.  Not sure how to deal with that.
-            snprintf(name, sizeof(name), "Local\\CYGWINX_WINDOWSDRI_%08x", draw->base.pDraw->id);
+            snprintf(name, sizeof(name), "Local\\CYGWINX_WINDOWSDRI_%08x", (unsigned int)draw->base.pDraw->id);
             ErrorF("shared memory region name %s\n", name);
 
             // Create a file mapping backed by the pagefile
